requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.11.10
python-multipart==0.0.12
httpx==0.28.1
python-dotenv==1.0.1
PyMySQL==1.1.1
